refactor: BoardStructStrategy 제거 및 구현 체 *Initializer로 명칭 변경
