+++ src/main/java/team/janggi/control/JanggiController.java
@@ -4,7 +4,7 @@ import java.util.Map;
 import team.janggi.domain.Board;
 import team.janggi.domain.Position;
 import team.janggi.domain.Team;
-import team.janggi.domain.strategy.boardstruct.NormalBoardStrategy;
+import team.janggi.domain.strategy.BoardInitializer;
 import team.janggi.domain.strategy.layout.normal.NormalLayoutStrategy;
 import team.janggi.domain.strategy.layout.normal.NormalSetup;
 import team.janggi.view.ConsoleInputView;
@@ -21,7 +21,7 @@ public class JanggiController {
 
     public void run() {
         final Board board = createBoard();
-        board.initBoard();
+        board.init();
 
         Team currentTurn = Team.CHO;
         while (true) {
@@ -34,7 +34,7 @@ public class JanggiController {
         final NormalSetup hanSetup = consoleInputView.readHanNormalSetup();
 
         final NormalLayoutStrategy layout = new NormalLayoutStrategy(choSetup, hanSetup);
-        final NormalBoardStrategy boardStructStrategy = new NormalBoardStrategy(layout);
+        final BoardInitializer boardStructStrategy = new BoardInitializer(layout);
         return new Board(boardStructStrategy);
     }
