+++ src/main/java/team/janggi/domain/strategy/BoardInitializer.java
@@ -1,17 +1,17 @@
-package team.janggi.domain.strategy.boardstruct;
+package team.janggi.domain.strategy;
 
 import team.janggi.domain.status.BoardStatus;
 import team.janggi.domain.Position;
 import team.janggi.domain.piece.Empty;
 import team.janggi.domain.strategy.layout.normal.NormalLayoutStrategy;
 
-public class NormalBoardStrategy implements BoardStructStrategy {
+public class BoardInitializer implements NormalBoardStructStrategy {
     private static final int NORMAL_BOARD_Y_SIZE = 10;
     private static final int NORMAL_BOARD_X_SIZE = 9;
 
     private final NormalLayoutStrategy layout;
 
-    public NormalBoardStrategy(NormalLayoutStrategy layout) {
+    public BoardInitializer(NormalLayoutStrategy layout) {
         this.layout = layout;
     }
