+++ src/test/java/team/janggi/domain/BoardTest.java
@@ -12,7 +12,7 @@ import team.janggi.domain.piece.Piece;
 import team.janggi.domain.piece.Soldier;
 import team.janggi.domain.status.BoardStateReader;
 import team.janggi.domain.status.LocalMemoryBoardStatus;
-import team.janggi.domain.strategy.boardstruct.NormalBoardStrategy;
+import team.janggi.domain.strategy.BoardInitializer;
 import team.janggi.domain.strategy.layout.normal.NormalLayoutStrategy;
 import team.janggi.domain.strategy.layout.normal.NormalSetup;
 
@@ -21,10 +21,10 @@ public class BoardTest {
     @Test
     void 기본_장기_보드_생성_테스트() {
         var layout = new NormalLayoutStrategy(NormalSetup.바깥상차림, NormalSetup.바깥상차림);
-        var boardStruct = new NormalBoardStrategy(layout);
+        var boardStruct = new BoardInitializer(layout);
 
         Board board = new Board(boardStruct);
-        board.initBoard();
+        board.init();
 
         Assertions.assertEquals(90, board.getStatus().size());
     }
@@ -32,10 +32,10 @@ public class BoardTest {
     @Test
     void 기본_장기_배치_테스트() {
         var layout = new NormalLayoutStrategy(NormalSetup.바깥상차림, NormalSetup.바깥상차림);
-        var boardStruct = new NormalBoardStrategy(layout);
+        var boardStruct = new BoardInitializer(layout);
 
         Board board = new Board(boardStruct);
-        board.initBoard();
+        board.init();
 
         BoardStateReader pieceMap = board.getStatus();
         // 초나라(CHO)가 아래쪽에 잘 배치되었는지 확인
@@ -96,10 +96,10 @@ public class BoardTest {
 
     private void runTest(NormalSetup choSetup, NormalSetup hanSetup) {
         var layout = new NormalLayoutStrategy(choSetup, hanSetup);
-        var boardStruct = new NormalBoardStrategy(layout);
+        var boardStruct = new BoardInitializer(layout);
 
         Board board = new Board(boardStruct);
-        board.initBoard();
+        board.init();
 
         BoardStateReader pieceMap = board.getStatus();
 
@@ -157,11 +157,11 @@ public class BoardTest {
 
     @Test
     public void 기물_이동_테스트() {
-        var boardStruct = new NormalBoardStrategy(EmptyLayoutStrategy.instance);
+        var boardStruct = new BoardInitializer(EmptyLayoutStrategy.instance);
         var boardStatus = new LocalMemoryBoardStatus();
         var board = new Board(boardStatus, boardStruct);
 
-        board.initBoard();
+        board.init();
 
         // 졸 기물 세팅
         Soldier soldier = new Soldier(Team.CHO);
