+++ src/test/java/team/janggi/domain/piece/GuardTest.java
@@ -9,7 +9,7 @@ import team.janggi.domain.EmptyLayoutStrategy;
 import team.janggi.domain.status.LocalMemoryBoardStatus;
 import team.janggi.domain.Position;
 import team.janggi.domain.Team;
-import team.janggi.domain.strategy.boardstruct.NormalBoardStrategy;
+import team.janggi.domain.strategy.BoardInitializer;
 
 public class GuardTest {
 
@@ -17,7 +17,7 @@ public class GuardTest {
 
     @BeforeEach
     void setUp() {
-        new NormalBoardStrategy(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
+        new BoardInitializer(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
     }
 
     @ParameterizedTest
