+++ src/test/java/team/janggi/domain/piece/SoldierTest.java
@@ -8,7 +8,7 @@ import team.janggi.domain.EmptyLayoutStrategy;
 import team.janggi.domain.status.LocalMemoryBoardStatus;
 import team.janggi.domain.Position;
 import team.janggi.domain.Team;
-import team.janggi.domain.strategy.boardstruct.NormalBoardStrategy;
+import team.janggi.domain.strategy.BoardInitializer;
 
 public class SoldierTest {
 
@@ -16,7 +16,7 @@ public class SoldierTest {
 
     @BeforeEach
     void setUp() {
-        new NormalBoardStrategy(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
+        new BoardInitializer(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
     }
 
     @Test
